requirements.txt
boto3==1.4.7
botocore==1.7.28
dj-database-url==0.4.1
Django==3.1.13
django-allauth==0.44
django-appconf==1.0.2
django-autofixture==0.12.1
django-compressor==2.1.1
django-core==1.4.1
django-csp==3.7
django-registration==3.1.2
django-rq==2.4.1
django-s3direct==0.4.2
django-storages==1.5.1
django-taggit==1.3.0
django-debreach==2.0.1
djangorestframework==3.12.2
docutils==0.14
gunicorn==19.6.0
jmespath==0.9.3
mailchimp==2.0.9
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-dateutil==2.6.1
pytz==2017.3
rcssmin==1.0.6
redis==3.5.3
rq==1.10.0
requests==2.20.0
rjsmin==1.0.12
s3transfer==0.1.11
simplejson==3.12.0
six==1.15.0
whitenoise==5.2.0
newrelic
